+++ internal/models/currency.go
@@ -8,4 +8,5 @@ type CurrencyShort struct {
 	Blockchain    *Blockchain `db:"blockchain" json:"blockchain"`
 	IsBitcoinLike bool        `db:"is_bitcoin_like" json:"is_bitcoin_like"`
 	IsEVMLike     bool        `db:"is_evm_like" json:"is_evm_like"`
+	IsStableCoin  bool        `db:"is_stable_coin" json:"is_stable_coin"`
 } // @name CurrencyShort